refactor(gfql): extract seeded fast-path specializations to dedicated modules (#1755)
Pure code move (no behavior change), following the gfql_fast_paths.py #1731
convention (keep the orchestrator readable; one-directional imports, no
back-edge). Moves the #1755 seeded typed-hop specializations out of the big
chain.py / gfql_unified.py orchestrators:

- NEW graphistry/compute/chain_fast_paths.py: _seeded_typed_hop_pandas_cudf,
  _seeded_typed_return_dst_pandas_cudf, _seeded_typed_return_dst_polars (moved
  verbatim; imports only leaves — .ast, .typing, Plottable).
- gfql_fast_paths.py (existing): gains _execute_seeded_typed_hop_fast_path.
- chain.py / gfql_unified.py: import the moved helpers (call sites unchanged).

Import edges (all one-directional, no cycles): chain -> chain_fast_paths;
gfql_unified -> gfql_fast_paths -> chain_fast_paths; chain_fast_paths -> leaves.
41 seeded tests pass unchanged; mypy clean on all four files.
